Moved shortcuts to own file and added view hotkey
